requirements.txt
opencv-python>=4.8.0
pyaudio>=0.2.13
grpcio>=1.59.0
torch>=2.1.0
mediapipe>=0.10.7
pygame>=2.5.2
openai>=1.3.0
pydub>=0.25.1
PyYAML>=6.0.1
protobuf>=4.24.4
yandexcloud>=0.1
numpy>=1.24.3